Add -reuse-unit flag to allow reusing an existing systemd cgroup/unit
This flag, which works only if specified in combination with the
-foreground switch causes docker to not create a new transient unit
for the container, but rather reuse the current systemd unit for
resource limits.

This allows e.g. a systemd unit file to contain:
ExecStart=docker run -f -reuse-unit the-image

And docker will use the cgroup of the foo.service unit, allowing
nice integration with systemd.

Docker-DCO-1.1-Signed-off-by: Alexander Larsson <alexl@redhat.com> (github: alexlarsson)

Conflicts:
	runconfig/parse.go