cgroup: Ensure --foreground mode automatically inherits the parent unit
This causes docker to not create a new transient unit for cgroup
containment and resource limits. Instead it will reuse the current
cgroup and apply resource limitations to it using SetUnitProperties.

This is interesting because it allows you to have a systemd unit file like:

[Service]
ExecStart=docker run --foreground run image
MemoryLimit=100M

With the following properties:

 * Any limits and other properties set up in the service file applies
   to both the docker client and the container "as one unit".
 * There is only one unit representing the service
 * You can kill the service via systemd which will kill both client and
   container.
 * If the client dies systemd will (first nicely then violently) kill the
   other processes in the cgroup (i.e. the container).

Since it is very problematic if you accidentally apply the device limits
on a user session we disallow the use of --foreground for user sessions.

Docker-DCO-1.1-Signed-off-by: Alexander Larsson <alexl@redhat.com> (github: alexlarsson)