Generate distribution-specific Dockerfile
This PR will generate a Dockerfile specific to a distribution.
Dockerfiles are generated for RHEL, CentOS and Fedora with this commit but
can be extended for other distributions.

This will allow distro-specific Dockerfile to be used in build for 'make test'.
It reads from /etc/redhat-release and generates a new Dockerfile for that
distribution.

Signed-off-by: Sally O'Malley <somalley@redhat.com>
Signed-off-by: Regan McCooey <rmccooey27@aol.com>

- adjust deps in Fedora.json
- add to fedora-1.8 and fedora-1.7 branches
- test-unit and test-docker-py successful

Signed-off-by: Lokesh Mandvekar <lsm5@fedoraproject.org>