Add --init=INITSYSTEM option to run the container in different init system mode
If you are running a container with an init system like systemd as pid 1 this i
option will modify the default container environment to follow the standard
defined for that init system.

Specifying --init=systemd will run the container with the following changes:

 /run mounted as a tmpfs
 /sys/fs/cgroup volume mounted from the host into the container read-only
 /var/log/journal/UUID volume mounted from the host into the container to allow
journald data from the container to be seen on the host.
 Changes the default signals send from docker stop to the init process (systemd)

Docker-DCO-1.1-Signed-off-by: Dan Walsh <dwalsh@redhat.com> (github: rhatdan)